% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-point-logitpart-const.R
\name{constLogitpart}
\alias{constLogitpart}
\title{Fitting Linear Models - Constat}
\usage{
constLogitpart(
  X1,
  X2,
  y,
  coef1,
  alphahat,
  from,
  to,
  by,
  tau,
  status = status,
  ylim = c(0, 2000),
  xlim = c(0, 100),
  plot = TRUE
)
}
\description{
Fitting Linear Models - Constat
}
