% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-comparison.R
\name{CoxSnell_BSmixture}
\alias{CoxSnell_BSmixture}
\title{Residuals - mixture Bernoulli/Birnbaum-Saunders}
\usage{
CoxSnell_BSmixture(x,X1,X2,theta,status,tau)
}
\arguments{
\item{x}{response variable.}

\item{X1}{model matrix.}

\item{X2}{model matrix.}

\item{theta}{estimative of the parameters.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{tau}{Is a prefixed limiting values. Default for ZERO.}
}
\description{
CoxSnell_BSmixture  is a function which extracts model residuals from Bernoulli/Birnbaum-Saunders model.
}
