% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-functions.R
\name{initialValuesBS}
\alias{initialValuesBS}
\title{Monte Carlo simulations}
\usage{
Lbs(X,y,status,tau=0,initialpoint,method="BFGS",iterations=10000)
}
\arguments{
\item{y}{shape parameter.}

\item{status}{Monte Carlo replications.}

\item{tau}{If is TRUE print the output.}
}
\description{
In this function, we evaluate the performance of the mixture model through MC simulations.
}
