% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-functions.R
\name{initialValuesLS}
\alias{initialValuesLS}
\title{Initial Values}
\usage{
initialValuesLS(y,X,status)
}
\arguments{
\item{y}{The response variable.}

\item{X}{The model matrix.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}
}
\description{
Initial values for the parameters to be optimizes over.
}
