% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo-simulations.R
\name{Lbs}
\alias{Lbs}
\title{Fitting Linear Models}
\usage{
Lbs(X,y,status,tau=0,initialpoint,method="BFGS",iterations=10000)
}
\arguments{
\item{X}{The model matrix.}

\item{y}{The response used.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{tau}{Is a prefixed limiting values. Default for ZERO.}

\item{initialpoint}{Initial values for the parameters to be optimized over.}

\item{method}{The method to be used. The default is BFGS.}

\item{iterations}{The maximum number of iterations. Default to 10000.}
}
\description{
Lbs is used to fit mixture model.
}
