% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-mixture-logit-birnbaum-saunders.R
\name{LbsProfile}
\alias{LbsProfile}
\title{Fitting Linear Models - Profile}
\usage{
LbsProfile(
  X1,
  X2,
  y,
  coef1,
  coef2,
  from,
  to,
  by,
  tau,
  status = status,
  ylim = c(0, 2000),
  xlim = c(0, 100),
  plot = TRUE
)
}
\description{
Fitting Linear Models - Profile
}
