% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo-simulations.R
\name{sim_mix_bs}
\alias{sim_mix_bs}
\title{Monte Carlo simulations}
\usage{
sim_mix_bs(n=100,alpha=0.1,nrep=5000,print=FALSE)
}
\arguments{
\item{n}{sample size.}

\item{alpha}{shape parameter.}

\item{nrep}{Monte Carlo replications.}

\item{print}{If is TRUE print the output.}
}
\description{
In this function, we evaluate the performance of the mixture model through MC simulations.
}
