% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-birnbaum-saunders.R
\name{tobitbs}
\alias{tobitbs}
\title{Fitting Linear Models - tobit bs}
\usage{
tobitbs(X,y,status,tau=0,k,initialpoint,method="BFGS",iterations=10000,hessian="TRUE",logtau="FALSE")
}
\arguments{
\item{X}{The model matrix.}

\item{y}{The response used.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{tau}{Is a prefixed limiting values. Default for ZERO.}

\item{k}{value.}

\item{initialpoint}{Initial values for the parameters to be optimized over.}

\item{method}{The method to be used. The default is BFGS.}

\item{iterations}{The maximum number of iterations. Default to 10000.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}

\item{logtau}{Logical. Default for FALSE.}
}
\description{
Lbs is used to fit mixture model.
}
