% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-comparison.R
\name{tobitResidualsBS}
\alias{tobitResidualsBS}
\title{Residuals - tobit bs}
\usage{
tobitResidualsBS(y,alphahat,muhat,status)
}
\arguments{
\item{y}{response variable.}

\item{alphahat}{estimative of the shape parameter.}

\item{muhat}{fitted values.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}
}
\description{
tobitResidualsBS is a function which extracts model residuals from tobit-bs.
}
