% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-comparison.R
\name{tobitResidualsEnvelopeBS}
\alias{tobitResidualsEnvelopeBS}
\title{Envelope - tobit Birnbaum-Saunders}
\usage{
tobitResidualsEnvelopeBS(X,y,muhat,alphahat,n,k,status,nboot = 19,alpha=0.05)
}
\arguments{
\item{X}{model matrix.}

\item{y}{response variable.}

\item{muhat}{fitted values.}

\item{alphahat}{estimative of the shape parameter.}

\item{n}{sample size.}

\item{k}{number.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{nboot}{number of bootrstrap replicates.}

\item{alpha}{Probability level for the interval.}

\item{model}{an object of class formula.}
}
\description{
This function calculates overall and pointwise confidence envelopes for a curve based on bootstrap replicates of the curve evaluated at a number of fixed points.
}
