% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-comparison.R
\name{tobitResidualsEnvelopeBSBernoulli}
\alias{tobitResidualsEnvelopeBSBernoulli}
\title{Envelope - mixture Bernoulli/Birnbaum-Saunders}
\usage{
tobitResidualsEnvelopeBSBernoulli(y,X1,X2,theta,status,tau,nboot = 19,alpha=0.05)
}
\arguments{
\item{y}{response variable.}

\item{X1}{model matrix.}

\item{X2}{model matrix.}

\item{theta}{estimative of the parameters.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{tau}{Is a prefixed limiting values. Default for ZERO.}

\item{nboot}{number of bootrstrap replicates.}

\item{alpha}{Confidence level for the interval.}
}
\description{
This function calculates overall and pointwise confidence envelopes for a curve based on bootstrap replicates of the curve evaluated at a number of fixed points.
}
