% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-comparison.R
\name{tobitResidualsNO}
\alias{tobitResidualsNO}
\title{Envelope - tobit normal}
\usage{
tobitResidualsNO(model,nboot = 19,alpha=0.01,tau=0.1,intercept = "TRUE")
}
\arguments{
\item{model}{an object of class formula.}

\item{nboot}{number of bootrstrap replicates.}

\item{alpha}{Confidence level for the interval.}

\item{tau}{Is a prefixed limiting values. Default for ZERO.}

\item{intercept}{logical. Should an intercept be included in the null model?}
}
\description{
This function calculates overall and pointwise confidence envelopes for a curve based on bootstrap replicates of the curve evaluated at a number of fixed points.
}
